Crate gobu

A crate for creating Visual Novels.

See example_vn for a good example of how to use this crate.

GameBuilder is recommended to make a Game with as you can use TOML files to make the game without having to use any of the functions.

Here is a quick example of how to quickly make a game.

extern crate gobu;
use gobu::{Game, GameBuilder};
let mut game: Game = GameBuilder::new([WIDTH, HEIGHT]) // The width and height of the game/screen.
                .gui("./path/to/gui.toml") // Config files are in the TOML format.
                .characters("./path/to/characters.toml") // See the example_vn repo for more.
                .backgrounds("./path/to/backgrounds.toml")
                .build(&mut factory); // Requires a GfxFactory to build the game as it uses texture.

Re-exports

pub use game::Game;

Modules

character
Use both Characters and CharacterEntitys to display people or objects.

error
A collection of all the errors used in this crate.

game
The module which uses all the other modules to make a game.

gui
Manages the UI, which consists of only the textbox for now.

images
Deals with displaying the background image

input
The Piston events are taken and turned into GameEvents

music
The music that is stored and played by the game.

script
Manages the story of the VN and also parsing script files.

util
A variety of utility functions that make life easier.

Structs

GameBuilder
The struct that is used to build a Game struct.

Rect
Represents a rectangle

Type Definitions

Pos
Represents a position as a vector of f64s