Trait gobu::animation::CharacterTransition

pub trait CharacterTransition {
    fn create(&self) -> Box<CharacterTransition>;

    fn update(&mut self, image: &mut Image, delta_time: f64) -> TransResult;

    fn finish(&mut self, image: &mut Image);
}

A trait for Character Transitions like FadeIn

Required Methods

fn create(&self) -> Box<CharacterTransition>

Create copy of itself. This is different to new as it requires a reference to self. This is because the struct which implements this trait shall be stored in a HashMap and therefore when a transition is needed we take the object in the HashMap and call [create] on it.

fn update(&mut self, image: &mut Image, delta_time: f64) -> TransResult

Every time the game updates, the entity will call this method and provide it's image. delta_time is the amount of time that has passed since the last update event. This will return a TransResult. If Finished then the animation shall be removed. If Continue nothing will happen.

fn finish(&mut self, image: &mut Image)

If the user wanted to skip the animation then this method will be called to make the character go to its final destination. For example if this was FadeIn, the image be fully opaque. If this was SlideIn, the image would be at its final destination.