Module gobu::character

Use both Characters and CharacterEntitys to display people or objects.

A [Character'] is a sort of template which holds all the different states and their corresponding textures. The characters are created first and then during the game they are spawned onto the stage as a [CharacterEntity`].

A CharacterEntity is the struct that is shown on screen. These hold positions but only have on state at a time. The state is changed by going back to the original Character and finding the desire state. Entities can have the same name as the Character but it is not recommended if you intend on having multiple entities of the same character in the stage at the same time.

Structs

Character
A character that is used to spawn an entity onto the stage.

CharacterEntity
The entity of a character that is spawned into the stage.