Struct gobu::character::Character

pub struct Character {
    pub default: String,
    pub state_map: HashMap<String, Arc<G2dTexture>>,
    pub size: [f64; 2],
    pub offset: Pos,
}

A character that is used to spawn an entity onto the stage.

Fields
default: String

The default state that the entity will have upon being spawned.

state_map: HashMap<String, Arc<G2dTexture>>

A collection of all the different states and their corresponding textures.

size: [f64; 2]

The size of the images [Width, Height]

offset: Pos

The offset (in percentage) of the centre of the images. This is helpful when you want to move the entity. x = 0., y = 0. : Top-Left corner x = 0.5, y = 0.5 : Centre of the image

Methods

impl Character

pub fn new(
default: String,
texture: Arc<G2dTexture>,
size: [f64; 2]
) -> Character

Create a new character where texture is the default texture that is to be used and default is the name of the default texture.

pub fn spawn(&self, name: String) -> Option<CharacterEntity>

Spawns an entity of the character onto stage with the given name.

pub fn add_state(&mut self, name: String, texture: Arc<G2dTexture>)

Adds a state to the character. This can be something like "sad" or "happy"

pub fn set_offset(&mut self, x: f64, y: f64)

Sets the offset of the character. This is (0, 0) by default.