Struct gobu::character::CharacterEntity

pub struct CharacterEntity {
    pub image: Image,
    pub texture: Arc<G2dTexture>,
    pub rect: Rect,
    pub visible: bool,
    pub name: String,
    pub offset: Pos,
}

The entity of a character that is spawned into the stage.

Fields
image: Image

The Image primitive provided by the Piston libraries.

texture: Arc<G2dTexture>

The current texture being used to draw the character.

rect: Rect

The position and size of the character image.

visible: bool

Whether the character is visible on screen or not.

name: String

The name of the entity.

offset: Pos

The offset of the images so that a specific point on the image can be used as the centre. Set to (0, 0) by default.

Methods

impl CharacterEntity

pub fn set_visible(&mut self, visible: bool)

Sets the visibility of the character on screen.

pub fn set_texture(&mut self, texture: Arc<G2dTexture>)

Sets the texture of the character.

pub fn set_pos(&mut self, pos: Pos)

Sets the position of the character. This accounts for the offset.

pub fn draw(&self, c: Context, g: &mut G2d)

Draws the character onto the screen.