Struct gobu::game::Game

pub struct Game {
    pub size: Rect,
    pub ui: Ui,
    pub background: BackgroundImage,
    pub backgrounds: HashMap<String, Arc<G2dTexture>>,
    pub stage: HashMap<String, CharacterEntity>,
    pub characters: HashMap<String, Character>,
    pub story: Script,
    pub grid: Grid,
    pub input: GameInput,
    pub music: Option<Music>,
}

The game

Fields
size: Rect

The rectangle that represents the game window

ui: Ui

The UI component which manages the textbox(s)

background: BackgroundImage

The background of the game

backgrounds: HashMap<String, Arc<G2dTexture>>

A HashMap of all the backgrounds that the background can be set to

stage: HashMap<String, CharacterEntity>

A HashMap of all the characters on the screen / in the world

characters: HashMap<String, Character>

A HashMap of all the characters that are or can be put onto the stage

story: Script

Holds a collection of all the scripts and also the current script

grid: Grid

A grid that allows you to set the position without having to bother with the actual height and width of the screen

input: GameInput

Holds all the possible inputs for the game and what events they produce

music: Option<Music>

The music that is played in the game. (This is optional)

Methods

impl Game

pub fn new(size: [f64; 2]) -> Game

Create a new Game struct by providing the [width, height] of the screen.

It is recommended to use GameBuilder to create your Game struct instead.

pub fn handle_event(&mut self, event: &Event)

Handles a piston event. Currently only going forward in the story is supported.

pub fn draw(&mut self, c: Context, g: &mut G2d)

Draws all the components of the game EXCEPT the text.

Use draw_text to set the text first.

pub fn draw_text(&mut self, brush: &mut GlyphBrush<Resources, GfxFactory>)

Slightly misleading as it doesn't draw it just yet, it just queues all the glyphs ready to be drawn.

To truly draw the text use the draw_2d_with_text function that is implemented on PistonWindow but remember to import the GfxGlyph trait first.

pub fn add_character(&mut self, name: String, character: Character)

Add a Character to the HashMap of characters

pub fn add_background(&mut self, name: String, background: Arc<G2dTexture>)

Add a texture to the HashMap of backgrounds

pub fn add_to_stage(&mut self, name: String, character: String) -> bool

Add a character to the stage using the character argument as the key to the characters HashMap and the name argument is the name of the CharacterEntity That is spawned.

Returns True if the character is spawned to the stage.

pub fn set_background(&mut self, name: &String)

Set the background using the name argument as the key to its texture found in the backgrounds HashMap.

pub fn change_entity_state(&mut self, name: &String, state: &String) -> bool

Change the state of a CharacterEntity on the stage. Returns True if the entity is changed.

pub fn move_character(&mut self, name: &String, pos: Pos)

Move a CharacterEntity's position on the screen. (The position is relative to the grid you've specified).

pub fn apply_grid(&mut self, w: u32, h: u32)

Apply a grid to the game to make it easier to position characters. W is how many cells make up the width of the screen and h is how many cells make up the height of the cell.

pub fn set_textbox(&mut self, text_box: TextBox)

Set the main textbox of the game.

pub fn set_speaker_box(&mut self, speaker_box: TextBox)

Set the speaker box which is the textbox that holds the name of the current speaker. Is not shown by default.

pub fn enable_music(&mut self) -> Result<&mut Music, MusicError>

Enables playing music in the game and returns a mutable reference in a result to the newly created Music struct that can be used to play music.

   if let Ok(music) = game.enable_music() {
       music.add_music_from_file("gabriel", "C:\\Users\\Hiruna\\Music\\Songs\\01. ガヴリールドロップキック.flac")
           .expect("Error in loading the music!");
       music.set_music("gabriel");
   }

pub fn load_characters_from_file<P: AsRef<Path>>(
&mut self,
path: P,
factory: &mut GfxFactory
) -> Result<(), ConfigImportError>

Load characters from a TOML file.

pub fn load_characters_from_str(
&mut self,
text: &str,
factory: &mut GfxFactory
) -> Result<(), ConfigImportError>

Load characters from a &str. Use this if you intend on compiling the TOML file and not keep it externally. Also could be used to load characters after decrypting a file.

pub fn load_input_from_file<P: AsRef<Path>>(
&mut self,
path: P
) -> Result<(), ConfigImportError>

Load input from a TOML file.

pub fn load_input_from_str(
&mut self,
text: &str
) -> Result<(), ConfigImportError>

Load input from a &str. Use this if you intend on compiling the TOML file and not keep it externally. Also could be used to load characters after decrypting a file.

pub fn load_gui_from_file<P: AsRef<Path>>(
&mut self,
path: P
) -> Result<(), ConfigImportError>

Load the GUI from a TOML file.

pub fn load_gui_from_str(&mut self, text: &str) -> Result<(), ConfigImportError>

Load the GUI from a &str. Use this if you intend on compiling the TOML file and not keep it externally. Also could be used to load characters after decrypting a file.

pub fn load_backgrounds_from_file<P: AsRef<Path>>(
&mut self,
path: P,
factory: &mut GfxFactory
) -> Result<(), ConfigImportError>

Load the backgrounds from a TOML file.

pub fn load_backgrounds_from_str(
&mut self,
text: &str,
factory: &mut GfxFactory
) -> Result<(), ConfigImportError>

Load the backgrounds from a &str. Use this if you intend on compiling the TOML file and not keep it externally. Also could be used to load characters after decrypting a file.

pub fn load_scripts_from_file<P: AsRef<Path>>(
&mut self,
path: P
) -> Result<(), ScriptConfigImportError>

Load the scripts from a TOML file.

pub fn load_scripts_from_str(
&mut self,
text: &str
) -> Result<(), ScriptConfigImportError>

Load scripts from a &str. Use this if you intend on compiling the TOML file and not keep it externally. Also could be used to load characters after decrypting a file.

impl Game

pub fn next_step(&mut self)

Load the next step of the script and execute the step.