Struct gobu::gui::Ui

pub struct Ui {
    pub textbox: TextBox,
    pub speaker_box: Option<TextBox>,
}

A struct that manages all the textboxes.

Fields
textbox: TextBox

The main textbox that displays the text.

speaker_box: Option<TextBox>

The textbox that displays the name of the speaker. If None then nothing will be shown.

Methods

impl Ui

pub fn new(canvas: Rect) -> Ui

Creates a new Ui struct.

pub fn draw(&mut self, c: Context, g: &mut G2d)

Draws the components that the Ui contains.

pub fn draw_text(&mut self, brush: &mut GlyphBrush<Resources, GfxFactory>)

Queues the text into the Brush that will be drawn with [draw_2d_with_text]