Struct gobu::gui::textbox::TextBox

pub struct TextBox {
    pub rect: Rectangle,
    pub inner: Rect,
    pub outer: Rect,
    pub text_pos: f64,
    pub padding: Padding,
    pub text_changed: bool,
    pub text: String,
    pub color: [f32; 4],
    pub font_scale: Scale,
}

A struct that represents and draws a textbox

Fields
rect: Rectangle

The rectangle which holds the x and y co-ordinates and the width and the height.

inner: Rect

The inner rectangle after applying padding. This is where the text will start.

outer: Rect

The outer rectangle which is what the square is drawn using.

text_pos: f64

The y position of the text.

padding: Padding

The padding applied on the textbox.

text_changed: bool

Whether the text has been changed or not.

text: String

The string that is to be drawn. gfx_glyph takes care of the wrapping.

color: [f32; 4]

The colour of the text.

font_scale: Scale

The size of the font.

Methods

impl TextBox

pub fn new(rect: Rect) -> TextBox

Create a new TextBox struct.

pub fn set_text(&mut self, text: String)

Sets the text that is shown by the textbox.

pub fn draw(&self, c: Context, g: &mut G2d)

Draw the textbox.

pub fn draw_text(&self, brush: &mut GlyphBrush<Resources, GfxFactory>)

Queue the text of the textbox into the brush.