Struct gobu::input::GameInput

pub struct GameInput {
    pub continue_: HashSet<Button>,
}

The struct in charge of handling the input of the game.

Fields
continue_: HashSet<Button>

the buttons that continue the story of the game.

Methods

impl GameInput

pub fn new() -> Self

Create a new GameInput struct

pub fn handle_event(&self, button: &Button) -> Option<GameEvent>

Handle a Piston Event and return a GameEvent.

pub fn add_continue_event(&mut self, button: Button)

Add an event that will continue the story.

pub fn add_continue_events(&mut self, buttons: HashSet<Button>)

Add events that will continue the story.

pub fn add_input(&mut self, input: GameInput)

Add input in the form of a GameInput struct. Just joins all of the data together.