Struct gobu::music::Music

pub struct Music {
    pub sink: Sink,
    pub library: HashMap<String, Arc<[u8]>>,
    pub loop_: bool,
}

The struct in charge of handling the music for the game.

Fields
sink: Sink

The Rodio Sink that music can be played in.

library: HashMap<String, Arc<[u8]>>

A collection of music data in bytes.

loop_: bool

Whether the music set will loop or not.

Methods

impl Music

pub fn new() -> Result<Self, MusicError>

Create a new Music struct. Returns an error if not possible.

pub fn set_music(&mut self, music: &str) -> Result<(), MusicError>

Sets the music that is to be played.

pub fn add_music_from_file<P: AsRef<Path>, S: ToString>(
&mut self,
name: S,
path: P
) -> Result<(), MusicError>

Add music from a file.

pub fn add_music(&mut self, name: String, music: Vec<u8>)

Add the buffered data of a music file.

pub fn set_loop(&mut self, loop_: bool)

If True will cause the music that is played to loop infinitely.

pub fn load_from_config_file<P: AsRef<Path>>(
&mut self,
path: P
) -> Result<(), ConfigImportError>

Loads music from a TOML file.

pub fn load_from_config_str(
&mut self,
text: &str
) -> Result<(), ConfigImportError>

Loads music from a [&str] using TOML syntax.