Enum gobu::script::ScriptStep

pub enum ScriptStep {
    Dialogue(String, String),
    DialogueContinue(String),
    Show(String, Option<String>),
    Hide(String),
    Spawn(String, Option<String>, Option<(f64, f64)>),
    Kill(String),
    Move(String, (f64, f64)),
    Stage(String),
    GoTo(Option<String>, Option<String>),
    End,
}

Represents an action dictated by the script

Variants
Dialogue(String, String)

A dialogue that consists of the name of the speaker and the text. (Name, Text)

DialogueContinue(String)

Only changes the text in the textbox. Doesn't change the speaker.

Show(String, Option<String>)

Shows an entity on the screen and if the state is not None will change that entity's state to the state named. (Entity, Some(State))

Hide(String)

Hides an entity using its name

Spawn(String, Option<String>, Option<(f64, f64)>)

Spawns a character with a specific entity name and also possibly a position. (CharacterName, Some(EntityName), Some(Position))

Kill(String)

Kill an entity

Move(String, (f64, f64))

Move an entity to a specific position

Stage(String)

Set the background

GoTo(Option<String>, Option<String>)

Go to a specific script or part of a script (ScriptName, AnchorName)

End

End of the script/game