Struct gobu::script::Script

pub struct Script {
    pub script: Vec<ScriptStep>,
    pub scripts: IndexMap<String, IndexMap<String, Vec<ScriptStep>>>,
    pub index: (usize, usize),
    // some fields omitted
}

The struct in charge of the story in the VN and also in charge of parsing everything.

Fields
script: Vec<ScriptStep>

The current script that is being used.

scripts: IndexMap<String, IndexMap<String, Vec<ScriptStep>>>

A collection of all the scripts that can be used.

index: (usize, usize)

The current index of what script is being used and what anchor is being used.

Methods

impl Script

pub fn new() -> Script

Create a new Script struct.

pub fn set_script(&mut self, name: &str, anchor: Option<String>)

Set a script with its name and maybe its anchor.

pub fn next_script(&mut self)

Go to the next script.

pub fn load_script(
&mut self,
name: String,
scripts: IndexMap<String, Vec<ScriptStep>>
)

Load a script into the collection of scripts.

pub fn load_scripts(
&mut self,
map: IndexMap<String, IndexMap<String, Vec<ScriptStep>>>
)

Load a number of scripts into the collection of scripts.

pub fn load_from_str(
&mut self,
name: String,
content: &str
) -> Result<(), ScriptImportError>

Load scripts from a [&str].

pub fn load_from_file<P: AsRef<Path>>(
&mut self,
name: String,
path: P
) -> Result<(), ScriptImportError>

Load scripts from an external file.