Struct gobu::GameBuilder

pub struct GameBuilder { /* fields omitted */ }

The struct that is used to build a Game struct.

Methods

impl GameBuilder

pub fn new(size: [f64; 2]) -> Self

Create a new GameBuilder struct by providing the size of the window.

pub fn build(self, factory: &mut GfxFactory) -> Result<Game, GameBuildError>

Builds the Game or returns an error.

pub fn gui<F: Into<ExtFile>>(self, file: F) -> Self

Create a GUI from either a path to an external file or a str.

pub fn characters<F: Into<ExtFile>>(self, file: F) -> Self

Create Characters from either a path to an external file or a str.

pub fn backgrounds<F: Into<ExtFile>>(self, file: F) -> Self

Create Backgrounds from either a path to an external file or a str.

pub fn input<F: Into<ExtFile>>(self, file: F) -> Self

Create Inputs from either a path to an external file or a str.

pub fn grid(self, w: u32, h: u32) -> Self

Create a Grid from either a path to an external file or a str.

pub fn scripts<F: Into<ExtFile>>(self, file: F) -> Self

Load scripts from either a path to an external file or a str.