Module gobu::util

A variety of utility functions that make life easier.

Contains specific functions that deal with the gfx_glyph crate but will only be compiled if the gfx_glyph_text feature is enabled.

Re-exports

pub use self::gfx_glyph::GfxGlyph;

Modules

gfx_glyph
A trait that implements the draw_2d_with_text function onto PistonWindow.

load
Helpful functions that load assets from a TOML file.

Structs

GameBuilder
The struct that is used to build a [Game] struct.

Enums

ExtFile
Represents an external file OR a String.

Functions

load_backgrounds_from_file
Load the backgrounds from a TOML file

load_characters_from_file
Load the characters from a TOML file

load_gui_from_file
Load the GUI from a TOML file

load_input_from_file
Load the inputs from a TOML file

load_scripts_from_file
Load the scripts from a TOML file