Struct gobu::gui::textbox::TextBox

pub struct TextBox {
    pub rect: Rectangle,
    pub inner: Rect,
    pub outer: Rect,
    pub text_pos: f64,
    pub padding: Padding,
    pub text_changed: bool,
    pub text_primitive: Text,
    pub font_size: u32,
    pub text_v: Vec<String>,
    pub line_gap: u32,
}

A struct that represents and draws a textbox

Fields
rect: Rectangle

The rectangle which holds the x and y co-ordinates and the width and the height.

inner: Rect

The inner rectangle after applying padding. This is where the text will start.

outer: Rect

The outer rectangle which is what the square is drawn using.

text_pos: f64

The y position of the text.

padding: Padding

The padding applied on the textbox.

text_changed: bool

Whether the text has been changed or not.

text_primitive: Text

The Piston text primitive that is used.

font_size: u32

The font size that is used with the Piston text primitive.

text_v: Vec<String>

The text that is to be displayed. One item of the Vec is a line.

line_gap: u32

The gap between the lines.

Methods

impl TextBox

pub fn new(rect: Rect) -> TextBox

Create a new TextBox struct.

pub fn set_text(&mut self, text: String)

Sets the text that is shown by the textbox.

pub fn draw(&mut self, c: Context, g: &mut G2d, glyph_cache: &mut Glyphs)

Draw the textbox.